feat: Adicionando findById

+++ internal/repository/fornecedorInterface.go
@@ -7,7 +7,7 @@ import (
 type FornecedorInterface interface {
 	Create(f *types.Fornecedor) error
 	FindAll() ([]*types.Fornecedor, error)
-	FindById(id uint) (*types.Fornecedor, error)
+	FindById(id uint64) (*types.Fornecedor, error)
 	IsUnique(cnpj string, instituicaoId uint64) (bool, error)
 	Update(f *types.Fornecedor) error
 	Delete(f *types.Fornecedor) error
